requirements.txt
streamlit==1.12.0
gtts==2.2.3
PyPDF2==3.0.0
python-docx==0.8.11
pydub==0.25.1